fix ios WebSocket issues

+++ Assets/ConnectApp/Utils/Sqlite/Models/DBMessageLite.cs
@@ -3,7 +3,7 @@ using SQLite4Unity3d;
 namespace ConnectApp.Utils {
     public class DBMessageLite {    
         [PrimaryKey]
-        public int nonce { get; set; }
+        public long nonce { get; set; }
         public string messageId { get; set; }
         public string channelId { get; set; }
